Minor package rename
