+++ logical-types/src/main/java/io/confluent/kafka/schemaregistry/type/logical/policy/CompatibilityResult.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.confluent.kafka.schemaregistry.type.logical.check;
+package io.confluent.kafka.schemaregistry.type.logical.policy;
 
 import java.util.ArrayList;
 import java.util.Collections;
