+++ logical-types/src/test/java/io/confluent/kafka/schemaregistry/type/logical/policy/LogicalTypeCheckerTest.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package io.confluent.kafka.schemaregistry.type.logical.check;
+package io.confluent.kafka.schemaregistry.type.logical.policy;
 
 import io.confluent.kafka.schemaregistry.type.logical.LogicalType;
 import io.confluent.kafka.schemaregistry.type.logical.Schema;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.confluent.kafka.schemaregistry.type.logical.check.LogicalTypeChecker.Mode;
+import io.confluent.kafka.schemaregistry.type.logical.policy.LogicalTypeChecker.Mode;
 import io.confluent.kafka.schemaregistry.type.logical.Schema.Field;
 
 import org.junit.jupiter.api.Test;
